# Avoid the use of shared variable.