- Use register instead.
- Fix Python benchmark script with latest pandas and XGBoost.

I ran the Python benchmark script using RTX8000 with CPU excluded:

**PR**
| model               | test_rows | cpu_time(s) | cpu_std | gpu_time(s)          | gpu_std               | speedup |
|---------------------|-----------|-------------|---------|----------------------|-----------------------|---------|
| adult-small         | 10000     | 0.0         |         | 0.011279034591279925 | 0.0071617705732274824 |         |
| adult-med           | 10000     | 0.0         |         | 0.11670323801226914  | 0.011867081750644454  |         |
| adult-large         | 10000     | 0.0         |         | 9.429356226988602    | 0.047229702183655395  |         |
| covtype-small       | 10000     | 0.0         |         | 0.01151795060141012  | 0.003561754665202217  |         |
| covtype-med         | 10000     | 0.0         |         | 0.6641728987917304   | 0.0032803489983452214 |         |
| covtype-large       | 10000     | 0.0         |         | 74.24905845939648    | 0.5906834845783823    |         |
| cal_housing-small   | 10000     | 0.0         |         | 0.006276315008290112 | 0.002279411050170562  |         |
| cal_housing-med     | 10000     | 0.0         |         | 0.10911295339465141  | 0.0012658547831405737 |         |
| cal_housing-large   | 10000     | 0.0         |         | 23.429275520995724   | 0.08867194402118761   |         |
| fashion_mnist-small | 10000     | 0.0         |         | 0.16429676760453732  | 0.06816532400946669   |         |
| fashion_mnist-med   | 10000     | 0.0         |         | 4.30540003181668     | 0.07100682231932201   |         |
| fashion_mnist-large | 10000     | 0.0         |         | 178.59231641321676   | 1.0623992933829536    |         |


**main**
| model               | test_rows | cpu_time(s) | cpu_std | gpu_time(s)          | gpu_std               | speedup |
|---------------------|-----------|-------------|---------|----------------------|-----------------------|---------|
| adult-small         | 10000     | 0.0         |         | 0.010922096599824727 | 0.006964535125901965  |         |
| adult-med           | 10000     | 0.0         |         | 0.11171154719777406  | 0.012917956737291719  |         |
| adult-large         | 10000     | 0.0         |         | 9.12171239000745     | 0.16260832207140813   |         |
| covtype-small       | 10000     | 0.0         |         | 0.011427397408988326 | 0.003081750313730569  |         |
| covtype-med         | 10000     | 0.0         |         | 0.6582016726024449   | 0.010074246049065491  |         |
| covtype-large       | 10000     | 0.0         |         | 73.67104159318842    | 0.6321216129573327    |         |
| cal_housing-small   | 10000     | 0.0         |         | 0.005807545990683139 | 0.00245267488361271   |         |
| cal_housing-med     | 10000     | 0.0         |         | 0.10978057580068708  | 0.0016856217806583914 |         |
| cal_housing-large   | 10000     | 0.0         |         | 23.66984584759921    | 0.10060247707348238   |         |
| fashion_mnist-small | 10000     | 0.0         |         | 0.1685630762251094   | 0.06951830276227733   |         |
| fashion_mnist-med   | 10000     | 0.0         |         | 4.44043413939653     | 0.06073396821402969   |         |
| fashion_mnist-large | 10000     | 0.0         |         | 183.3073539715959    | 0.6199842390017297    |         |