@trivialfis 2024-08-19
@hcho3 Have you seen this error using the rapids test matrix generator:
```
  File "/opt/conda/lib/python3.11/site-packages/rapids_dependency_file_generator/_rapids_dependency_file_generator.py", line 421, in make_dependency_files
    raise ValueError(f"No matching matrix found in '{include}' for: {matrix_combo}")
ValueError: No matching matrix found in 'cuda_version' for: {'cuda': '12.5', 'arch': 'x86_64', 'py': '3.11'}
```
with `rapidsai/shared-workflows/.github/workflows/conda-cpp-tests.yaml`?