@bdice 2024-08-19
The failure on this PR is due to the addition of CUDA 12.5 support. This repository was missed during the upgrade. https://github.com/rapidsai/build-planning/issues/73

I went ahead and pushed the changes you need in 444005e3e2c21b9d87526d114c44989754704f40.