@bdice 2024-08-19
I also pushed commit e04ea9b288a9691ce47383862bb3b15dee92f060 with the changes from `ci/release/update-version.sh 24.10.00`. This repo is updated pretty irregularly and we need to keep all the RAPIDS versions aligned (not just the CI workflow branches).